Migration of orphaned Puppetclasses
When the Host or Hostgroup has no Environment associated, the Facet wont
get created and thus the following migrations for any Puppet info will
get skipped.

For Hostgroup, we create the Facet as environment doesn't need to be
set, but for Host that is invalid state and we clean up the other data.