Fix the taxable_taxonomies creation
When the migration creates the records in taxable_taxonomies, it does
not set any value to created_at and updated_at magic columns. It uses
`insert_all` which skips callbacks, so they remain null. There are
databases which enforces the column not to be null, therefore we need to
be setting the values manually.